API change to a more OOP-approach. Lays the ground for more configurations; Major refactoring and according changes to tests. WIP
