+++ .gitignore
@@ -50,4 +50,4 @@ lammps-stable*
 ovito-basic*
 
 #vmd
-vmd-python*
+vmd-python/
