+++ sasa_lammps/__init__.py
@@ -1,3 +1,3 @@
-from sasa_lammps.sasa_main import sasa
+from sasa_lammps.sasa_main import Sasa
 
-__all__ = ["sasa"]
+__all__ = ["Sasa"]
