+++ sasa_lammps/sasa_main.py
@@ -16,10 +16,13 @@ import signal
 
 from multiprocessing import Pool
 
-from sasa_lammps.gro2lammps import gro2lammps
-from sasa_lammps.sasa_core import create_sasa_xyz
+from sasa_lammps.sasa_core import (
+    create_sasa_xyz,
+    rotate_probe,
+    neighbor_finder,
+)
 from sasa_lammps.lammps_manager import get_lammps_executable
-from sasa_lammps.helper import (
+from sasa_lammps.utils import (
     check_files,
     count_atoms_in_mol,
     count_atoms_in_macromol,
@@ -27,9 +30,9 @@ from sasa_lammps.helper import (
     read_last_two,
 )
 from sasa_lammps.conversion import (
-    rotate_probe,
-    convert_data_file,
-    neighbor_finder,
+    Converter,
+    LammpsToXyzStrategy,
+    GromacsToLammpsStrategy
 )
 from sasa_lammps.postprocessing import (
     residue_analysis,
@@ -40,45 +43,132 @@ from sasa_lammps.postprocessing import (
     atom_analysis_plot
 )
 from sasa_lammps.constants import (
-    ELEM_LIBRARY,
     FF_PARAMS,
     DUMP_COM,
     SPEC,
-    RESIDUELIST,
+    RESIDUE_LIST,
     IN_PRE,
     IN_TEMPLATE,
+    DATA_FILE,
     KCAL_TO_EV
 )
 
 class Sasa:
+    """
+    Class to perform a computation of the SASA, as well as hold the information of its results.
+    Several files need to be provided in the construction of this object. Then, a computation can be performed with
+    ```
+    sasa.compute(*args)
+    ```
+    and a subsequent post-processing with
+    ´``
+    sasa.postprocess(*args)
+    ```
+    """
     def __init__(
         self,
-        gro_file,
-        data_file,
-        mol_file,
-        ff_str,
-        dump_str,
-        lammps_exe=None,
-        n_procs=1,
-        srad=1.4,
-        samples=100,
-        path=".",
+        gro_file: str,
+        mol_file: str,
+        ff_str: str,
+        dump_str: str,
+        lammps_exe: str = None,
     ):
+        """Constructor for the `Sasa` object.
+
+        Parameters
+        ----------
+        gro_file : str
+            Name of the gromacs file of the macromolecule
+        mol_file : str
+            Name of the LAMMPS mol file to use as probe of the SAS (solvent acessible surface)
+        ff_str : str
+            Force field parameters to provide to LAMMPS. See examples directory
+            https://docs.lammps.org/pair_style.html
+            https://docs.lammps.org/pair_coeff.html
+            Care must be taken because currently the 'unit real' in the in.template basically restricts to only use pair_style reaxff.
+        dump_str : str
+            Dump command to provide to LAMMPS. See examples directory
+            https://docs.lammps.org/dump.html
+        lammps_exe : str,
+            Path to the LAMMPS executable to be used. If None, an executable is downloaded. 
+        """
         self.gro_file = gro_file
-        self.data_file = data_file
         self.mol_file = mol_file
         self.ff_str = ff_str
         self.dump_str = dump_str
         self.lammps_exe = lammps_exe if lammps_exe is not None else get_lammps_executable()
+        self.data_file = DATA_FILE
+        
+    def compute(
+        self,
+        n_procs: int = 1,
+        srad: float = 1.4,
+        samples: int = 100,
+        path: str = ".", # os.getcwd() ?
+    ) -> None:
+        """
+        Compute the solvet accessible surface analysis.
+
+        Parameters
+        ----------
+        n_procs : int, optional
+            Number of LAMMPS instances to run in parallel (Default: 1)
+        srad : float, optional
+            Probe radius: Effectively a scaling factor for the vdW radii
+            (Default: 1.4, which is the most commonly used because its approx. the
+            radius of water)
+        samples : int, optional
+            Maximum points on the atomic vdW sphere to generate per atom (Default: 100)
+        path : str, optional
+            Execution path (Default: .)
+
+        Returns
+        -------
+        None
+        """
         self.n_procs = n_procs
         self.srad = srad
         self.samples = samples
         self.path = path
-        self.parent_pid = os.getpid()   # pid of the parent process: needed to kill in case of an exception
+        self.parent_pid = os.getpid()   # PID of the parent process: required in order to kill in case of an exception
 
-    def _process(self):
+        self._initialize()
+
+        # here starts the computation
+        self._exec_lammps_iterations(self.n_probes, self.e_mol, self.e_prob)
+
+        return 0
+    
+    def postprocess(self) -> None:
+        """
+        Perform postprocessing on the results.
+        """
+        # Todo:
+        # add checks that all necessary files exist
+        # provide methods to load result data, such that a postprocessing can be performed without running the computation
+
+        # atom analysis
+        neighbor = neighbor_analysis(self.path, SPEC, self.gro_file)
+        atom_analysis_result = atom_analysis(self.path, SPEC, neighbor)
+        atom_analysis_plot(self.path, neighbor, atom_analysis_result)
+
+        # residue analysis
+        residuelist(self.path, self.gro_file)
+        residue_analysis_result = residue_analysis(self.path, SPEC, RESIDUE_LIST)
+        residue_analysis_plot(self.path, residue_analysis_result)
+
+        # one could add a config-method to define more details about the postprocessing
+        # or at least allow arguments here to do so.
+        # Also, the postprocessing could be refactored to be more general.
+
+        return 0
+
+    def _initialize(self) -> None:
+        """
+        Perform a number of initialization steps to prepare the actual computation.
+        """
         # generate lammps data file
-        gro2lammps(self.path, ELEM_LIBRARY).convert(self.gro_file, self.data_file)
+        Converter(GromacsToLammpsStrategy()).convert(self.path, self.gro_file)
 
         # remove existing files and copy input templates...
         check_files(self.path)
@@ -88,35 +178,21 @@ class Sasa:
         write_params_file(self.dump_str, DUMP_COM)
 
         # get the energies for the isolated macro- and probe molecule, respectively
-        e_mol, e_prob = self._pre_calc()
+        self.e_mol, self.e_prob = self._pre_calc()
 
         # convert data file
-        self.xyz_file = convert_data_file(self.path, self.data_file)
+        self.xyz_file = Converter(LammpsToXyzStrategy()).convert(self.path, DATA_FILE)
 
         # create sasa position file
         self.sasa_positions = create_sasa_xyz(
             self.path, self.xyz_file, self.srad, self.samples
         )
-        n_probes = len(self.sasa_positions)  # need this only to get the total num of iterations
+        self.n_probes = len(self.sasa_positions)  # need this only to get the total num of iterations
 
         # build neigbor list
         self.neighbors = neighbor_finder(
-            self.path, self.data_file, self.sasa_positions
+            self.path, DATA_FILE, self.sasa_positions
         )
-
-        # execute
-        self._exec_lammps_iterations(n_probes, e_mol, e_prob)
-
-        # postprocessing
-        ## atom analysis
-        neighbor = neighbor_analysis(self.path, SPEC, self.gro_file)
-        result = atom_analysis(self.path, SPEC, neighbor)
-        atom_analysis_plot(self.path, neighbor, result)
-
-        ## residue analysis
-        residuelist(self.path, self.gro_file)
-        result = residue_analysis(self.path, SPEC, RESIDUELIST)
-        residue_analysis_plot(self.path, result)
 
         return 0
 
@@ -130,9 +206,7 @@ class Sasa:
             Energy of the macro molecule in kcal/mole
         e_prob : float
             Energy of the probe molecule in kcal/mole
-
         """
-
         self._run_lmp(
             IN_PRE,
             [0.0, 0.0, 0.0],
@@ -144,10 +218,24 @@ class Sasa:
         return e_mol, e_prob
 
     def _exec_lammps_iterations(self, n_probes: int, e_mol: float, e_prob: float) -> None:
-        """Execute LAMMPS singlepoints on SASA coordinates using a N-atomic probe"""
-
+        """
+        Execute LAMMPS singlepoints on SASA coordinates using a N-atomic probe
+        
+        Parameters
+        ----------
+        n_probes : int
+            Nunmber of probe points on the SAS
+        e_mol : float
+            Energy of the macro molecule in kcal/mole
+        e_prob : float
+            Energy of the probe molecule in kcal/mole
+        
+        Returns
+        -------
+        None
+        """
         # Count atoms in macro molecule
-        atom_number = count_atoms_in_macromol(Path(self.path) / self.data_file)
+        atom_number = count_atoms_in_macromol(Path(self.path) / DATA_FILE)
 
         # create final output file header and write to spec.xyz
         header = f"{n_probes}\natom\tx\ty\tz\tres\tetot/eV\teint/eV\n"
@@ -157,7 +245,7 @@ class Sasa:
         # rotate the probe molecule for n-atomic probes (n > 1)
         if count_atoms_in_mol(Path(self.path) / self.mol_file) > 1:
             self.rotations = rotate_probe(
-                self.path, self.data_file, self.sasa_positions, self.neighbors
+                self.path, DATA_FILE, self.sasa_positions, self.neighbors
             )
         else:
             self.rotations = np.zeros((n_probes, 4))
@@ -233,7 +321,7 @@ class Sasa:
 
         cmd = f"""
             mpirun -np 1 {self.lammps_exe} -in {in_file} \
-            -var DataFile {self.data_file} -var MolFile {self.mol_file} \
+            -var DataFile {DATA_FILE} -var MolFile {self.mol_file} \
             -var sasaX {pos[0]:.3f} -var sasaY {pos[1]:.3f} \
             -var sasaZ {pos[2]:.3f} -var rotAng {rot[0]:.3f} \
             -var rotVecX {rot[1]:.3f} -var rotVecY {rot[2]:.3f} \
@@ -254,81 +342,6 @@ class Sasa:
             pass
 
         return 0
-
-
-def sasa(
-    gro_file,
-    data_file,
-    mol_file,
-    ff_str,
-    dump_str,
-    lammps_exe=None,
-    n_procs=1,
-    srad=1.4,
-    samples=100,
-    path=".",
-):
-    """
-    Run the SASA (solvet accessible surface analysis) on a given macromolecule 
-    using a given probe molecule.
-    The package was designed to start from a gromacs file of the macromolecule.
-    For good simulation practices the macromolecule should be pre-equilibrated in water.
-    Care must be taken for N-atomic probe molecules: The package does not identify
-    a plane or something in the probe molecule. It just makes sure that at every
-    interaction site the probe faces the macromolecule with the same orientation.
-    However, the orientation itself is purely determined by the configuration
-    given in the mol_file.
-
-    Parameters
-    ----------
-    gro_file : str
-        Name of the gromacs file of the macromolecule
-    data_file : str
-        Name of the LAMMPS data file of the macromolecule
-    mol_file : str
-        Name of the LAMMPS mol file to use as probe of the SAS (solvent acessible surface)
-    ff_str : str
-        Force field parameters to provide to LAMMPS. See examples directory
-        https://docs.lammps.org/pair_style.html
-        https://docs.lammps.org/pair_coeff.html
-        Care must be taken because currently the 'unit real' in the in.template basically restricts to only use pair_style reaxff.
-    dump_str : str
-        Dump command to provide to LAMMPS. See examples directory
-        https://docs.lammps.org/dump.html
-    lammps_exe : str, optional
-        Full path to the LAMMPS executable. If not provided, will automatically
-        download and use pre-built LAMMPS binaries from https://download.lammps.org/static/
-    n_procs : int, optional
-        Number of LAMMPS instances to run in parallel (Default: 1)
-    srad : float, optional
-        Probe radius: Effectively a scaling factor for the vdW radii
-        (Default: 1.4, which is the most commonly used because its approx. the
-        radius of water)
-    samples : int, optional
-        Maximum points on the atomic vdW sphere to generate per atom (Default: 100)
-    path : str, optional
-        Execution path (Default: .)
-
-    Returns
-    -------
-    None
-    """
-
-    S = Sasa(
-        gro_file,
-        data_file,
-        mol_file,
-        ff_str,
-        dump_str,
-        lammps_exe,
-        n_procs,
-        srad,
-        samples,
-        path,
-    )
-    S._process()
-
-    return 0
 
 
 def main():
